refactor: 줄간격, 띄어쓰기, import 정리
