+++ src/main/java/calculator/domain/StringCalculator.java
@@ -1,9 +1,6 @@
 package calculator.domain;
 
 public class StringCalculator {
-
-    public static final String INVALID_VALUE_INCLUDED = "음수나 숫자 이외의 값이 포함되어 있습니다.";
-
     public static int splitAndSum(String expression) {
         if (expression.isBlank()) {
             return 0;
