+++ src/test/java/calculator/domain/NumbersTest.java
@@ -11,9 +11,9 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class NumbersTest {
     @Test
     @DisplayName("합계를 계산하는 부분")
-    void calculateSumTest(){
+    void calculateSumTest() {
         List<Number> numbers = new ArrayList<Number>();
-        for(int i = 1; i <= 10; i++) {
+        for (int i = 1; i <= 10; i++) {
             numbers.add(new Number(Integer.toString(i)));
         }
         assertThat(new Numbers(numbers).calculateSum()).isEqualTo(55);
